添加Lucene框架，实现站内搜素

+++ src/main/java/com/neal/myblog/service/ArticleByVisitorService.java
@@ -53,7 +53,7 @@ public interface ArticleByVisitorService {
      * @param articleId 文章ID
      * @return 包装类
      */
-    TArticleVO getArticleById(int articleId);
+    TArticleVO getArticleById(long articleId);
 
     /**
      * 用于查询上下页文章ID,以及文章标题
@@ -61,5 +61,5 @@ public interface ArticleByVisitorService {
      * @param articleId 文章ID
      * @return TArticleEX
      */
-    TArticleEX getArticlePreAndNext(int articleId);
+    TArticleEX getArticlePreAndNext(long articleId);
 }
